+++ packages/components/src/MultiSelect/MultiSelect.stories.js
@@ -39,6 +39,24 @@ class Photos extends React.Component {
 						/>
 					</div>
 					<div className="form-group">
+						<label className="control-label" htmlFor="storybook">
+							Empty
+						</label>
+						<MultiSelect id="storybook" options={[]} isLoading={this.state.loading} withCreateNew />
+					</div>
+					<div className="form-group">
+						<label className="control-label" htmlFor="storybook">
+							2 items
+						</label>
+						<MultiSelect
+							id="storybook"
+							options={this.state.photos?.splice(0, 2)}
+							isLoading={this.state.loading}
+							withCreateNew
+						/>
+					</div>
+
+					<div className="form-group">
 						{/* eslint-disable-next-line jsx-a11y/label-has-associated-control */}
 						<label className="control-label" htmlFor="useless">
 							another