+++ packages/faceted-search/stories/facetedSearch.stories.js
@@ -501,10 +501,10 @@ export const BasicSearchWithSliderPopin = {
 
 export const WithQuickSearchFilter = () => (
 	<FacetedSearch.Faceted id="my-faceted-search">
-		<p>Quick search will only suggest facets matching input (Connection name, author)</p>
+		<p>Quick search will only suggest facets matching input (Connection name, Author)</p>
 		<br />
 		<FacetedSearch.BasicSearch
-			quickSearchFacetsFilter={(term, facet) => facet.properties.label.includes(term)}
+			quickSearchFacetsFilter={(term, facets) => facets.filter(facet => facet.properties.label.includes(term))}
 			badgesDefinitions={[badgeAuthor, badgeName, badgeConnectionName]}
 			onSubmit={action('onSubmit')}
 			callbacks={callbacks}